Move most matching to its own module
